requirements.txt
absl-py==1.0.0
setuptools==49.6.0
numpy==1.21.5
flax==0.4.0
optax==0.1.1
matplotlib==3.5.1
jax==0.3.1
jaxlib==0.3.0
ott-jax==0.2.2
wandb==0.12.11
ml-collections==0.1.1